requirements.txt
tkinter
biopython
os
socket